Upgrading dynamodb journal plugin to Akka 2.4
This project had a dependency on spray-aws, which had a dependency on
akka 2.3 and spray.  Spray is no longer supported moving forward, so I
changed the dependency to only require the aws sdk.

There are some substantial changes in Akka 2.4 for persistence, most
notably:
- WriteConfirmations go away, so there is no more concept of imperanent
  deletes.  A lot of code around write confirmations was removed from
the plugin
- The asyncWriteMessages signature changed.  Instead of taking a
  Seq[PersistentRepr], it instead takes a Seq[AtomicWrite].  There are
also certain guarantees around the handling of AtomicWrites that must
be implemented; namely that every AtomicWrite has an associated
Try[Unit] that is returned.  This means that every AtomicWrite should
yield a successful result.  I had to change the way we handled
asyncWriteMessages to always ensure a response.

I beleive that this plugin will be backward compatible, as I did not
alter the way the items are being built.

I created an `integration-test.sh` that will run the tests.  One of the
tests was pulled from Martin Krasser's journal plugin; the other test
uses the JournalSpec.

I did not due the JournalPerfSpec as I am only running locally, and I
was not able to get the throughput hitting the amazondb local.