The Big Restructure #14
Changed the table structure to partition on seqNr/100 and store the
remainder in a range key. This means that only when writing elements
with Sort=0 we update the highest sequence number, saving 99 writes per
100 events. This makes reading the highest sequence number a bit slower,
but recovery should be a lot less frequent than writing an event.

This commit also sequences the writes performed as part of a single
WriteMessages command, fixing the previously inconsistent updates that
resulted from running all AtomicWrites in parallel.

fixes #14