Add setSelect(ALL_ATTRIBUTES) to retrieve snapshots
By doing so, the indexes don't need to have ProjectionStrategy ALL,
but KEYS_ONLY suffices. Thus, a duplicate of the payload doesn't need
to be stored in the index, which is more space efficient.

This allows to make the choice between time and space efficiency,
by selecting the projection strategy for the created index.
Depending on the choice, setting this flag for retrieval will make
either strategy work as efficient as intended.